add login to user state, fix eslinter add axios config
