+++ src/actions/favouriteTypes.js
@@ -1,0 +1,9 @@
+export const FETCH_FAVOURITES_REQUEST = 'FETCH_FAVOURITES_REQUEST';
+export const FETCH_FAVOURITES_SUCCESS = 'FETCH_FAVOURITES_SUCCESS';
+export const FETCH_FAVOURITES_FAILURE = 'FETCH_FAVOURITES_FAILURE';
+export const CREATE_FAVOURITE_REQUEST = 'CREATE_FAVOURITE_REQUEST';
+export const CREATE_FAVOURITE_SUCCESS = 'CREATE_FAVOURITE_SUCCESS';
+export const CREATE_FAVOURITE_FAILURE = 'CREATE_FAVOURITE_FAILURE';
+export const DELETE_FAVOURITE_REQUEST = 'DELETE_FAVOURITE_REQUEST';
+export const DELETE_FAVOURITE_SUCCESS = 'DELETE_FAVOURITE_SUCCESS';
+export const DELETE_FAVOURITE_FAILURE = 'DELETE_FAVOURITE_FAILURE';
