+++ src/actions/index.js
@@ -1,24 +1,9 @@
 import axios from 'axios';
 
 import {
-  FETCH_USERS_FAILURE,
-  FETCH_USERS_REQUEST,
-  FETCH_USERS_SUCCESS,
-  CREATE_USERS_FAILURE,
-  CREATE_USERS_REQUEST,
-  CREATE_USERS_SUCCESS,
   FETCH_MOTOS_FAILURE,
   FETCH_MOTOS_REQUEST,
   FETCH_MOTOS_SUCCESS,
-  FETCH_FAVOURITES_FAILURE,
-  FETCH_FAVOURITES_REQUEST,
-  FETCH_FAVOURITES_SUCCESS,
-  CREATE_FAVOURITE_FAILURE,
-  CREATE_FAVOURITE_REQUEST,
-  CREATE_FAVOURITE_SUCCESS,
-  DELETE_FAVOURITE_FAILURE,
-  DELETE_FAVOURITE_REQUEST,
-  DELETE_FAVOURITE_SUCCESS,
   FETCH_BOOKMOTO_FAILURE,
   FETCH_BOOKMOTO_REQUEST,
   FETCH_BOOKMOTO_SUCCESS,
@@ -27,34 +12,6 @@ import {
   CREATE_BOOKMOTO_SUCCESS,
 
 } from './types';
-
-const fetchUsersRequest = () => ({
-  type: FETCH_USERS_REQUEST,
-});
-
-const fetchUsersSuccess = users => ({
-  type: FETCH_USERS_SUCCESS,
-  payload: users,
-});
-
-const fetchUsersFailure = error => ({
-  type: FETCH_USERS_FAILURE,
-  payload: error,
-});
-
-const createUsersRequest = () => ({
-  type: CREATE_USERS_REQUEST,
-});
-
-export const createUsersSuccess = user => ({
-  type: CREATE_USERS_SUCCESS,
-  payload: user,
-});
-
-const createUsersFailure = error => ({
-  type: CREATE_USERS_FAILURE,
-  payload: error,
-});
 
 const fetchMotosRequest = () => ({
   type: FETCH_MOTOS_REQUEST,
@@ -67,48 +24,6 @@ const fetchMotosSuccess = motos => ({
 
 const fetchMotosFailure = error => ({
   type: FETCH_MOTOS_FAILURE,
-  payload: error,
-});
-
-const fetchFavouritesRequest = () => ({
-  type: FETCH_FAVOURITES_REQUEST,
-});
-
-const fetchFavouritesSuccess = motos => ({
-  type: FETCH_FAVOURITES_SUCCESS,
-  payload: motos,
-});
-
-const fetchFavouritesFailure = error => ({
-  type: FETCH_FAVOURITES_FAILURE,
-  payload: error,
-});
-
-const createFavouriteRequest = () => ({
-  type: CREATE_FAVOURITE_REQUEST,
-});
-
-const createFavouriteSuccess = moto => ({
-  type: CREATE_FAVOURITE_SUCCESS,
-  payload: moto,
-});
-
-const createFavouriteFailure = error => ({
-  type: CREATE_FAVOURITE_FAILURE,
-  payload: error,
-});
-
-const deleteFavouriteRequest = () => ({
-  type: DELETE_FAVOURITE_REQUEST,
-});
-
-const deleteFavouriteSuccess = moto => ({
-  type: DELETE_FAVOURITE_SUCCESS,
-  payload: moto,
-});
-
-const deleteFavouriteFailure = error => ({
-  type: DELETE_FAVOURITE_FAILURE,
   payload: error,
 });
 
@@ -140,63 +55,11 @@ const createBookMotoFailure = error => ({
   payload: error,
 });
 
-export const fetchUsers = (email, password) => dispatch => {
-  dispatch(fetchUsersRequest);
-  
-  axios.post('/login',{
-
-     email,
-     password
-    }
-    )
-    .then(response => {
-      const user = response.data;
-      console.log(user)
-
-      localStorage.setItem('motoToken', user.token);
-
-      if (user) {
-        dispatch(fetchUsersSuccess(user));
-        window.location.href = '/motorcycles';
-      } else {
-        dispatch(fetchUsersFailure('user do not exist'));
-      }
-    })
-    .catch(error => {
-      dispatch(fetchUsersFailure(error.message));
-    });
-};
-
-export const createUsers = (name, email, password) => dispatch => {
-  dispatch(createUsersRequest);
-  
-  axios.post('/register',
-    {
-      name,
-      email,
-      password
-    })
-    .then(response => {
-      const user = response.data;
-      localStorage.setItem('motoToken', user.token);
-      if (user) {
-        dispatch(createUsersSuccess(response.data));
-        window.location.href = '/motorcycles';
-      } else {
-        dispatch(createUsersFailure('This email is in use, write another email!'));
-      }
-    })
-    .catch(error => {
-      dispatch(createUsersFailure(error.message));
-    });
-};
-
 export const fetchMotos = () => dispatch => {
   dispatch(fetchMotosRequest);
   axios.get('/motocycles')
     .then(response => {
       const motos = response.data;
-      
 
       dispatch(fetchMotosSuccess(motos));
     })
@@ -205,58 +68,10 @@ export const fetchMotos = () => dispatch => {
     });
 };
 
-export const fetchFavourites = () => dispatch => {
-  dispatch(fetchFavouritesRequest);
-  
-  
-  axios.get(`/favourites`,{ mode: 'cors' })
-    .then(response => {
-      const motos = response.data;
-      console.log(motos)
-      dispatch(fetchFavouritesSuccess(motos));
-    })
-    .catch(error => {
-      dispatch(fetchFavouritesFailure(error.message));
-    });
-};
-
-export const createFavourite = (motoid) => dispatch => {
-  dispatch(createFavouriteRequest);
-  axios.post(`/favourites`,
-    {
-      motocycle_id: motoid,
-    })
-    .then(response => {
-      const motosf = response.data;
-      dispatch(createFavouriteSuccess(motosf));
-    })
-    .catch(error => {
-      dispatch(createFavouriteFailure(error.message));
-    });
-};
-
-export const deleteFavourite = (motoid) => dispatch => {
-  dispatch(deleteFavouriteRequest);
-  console.log(motoid)
-  axios.delete(`/favourites/${motoid}`,
-  {
-    motocycle_id: motoid,
-  })
-    .then(response => {
-      const favourites = response.data;
-
-      dispatch(deleteFavouriteSuccess(favourites));
-    })
-    .catch(error => {
-      dispatch(deleteFavouriteFailure(error.message));
-    });
-};
-
 export const fetchMotoBook = () => dispatch => {
   dispatch(fetchBookMotoRequest);
-  
-  
-  axios.get(`/tests`,{mode: 'cors'})
+
+  axios.get('/tests', { mode: 'cors' })
     .then(response => {
       const bookmotos = response.data;
 
@@ -269,7 +84,7 @@ export const fetchMotoBook = () => dispatch => {
 
 export const createMotoBook = (motoid, date, city) => dispatch => {
   dispatch(createBookMotoRequest);
-  axios.post(`/tests`,
+  axios.post('/tests',
     {
       motocycle_id: motoid,
       testDate: date,
@@ -282,8 +97,4 @@ export const createMotoBook = (motoid, date, city) => dispatch => {
     .catch(error => {
       dispatch(createBookMotoFailure(error.message));
     });
-};
-
-export const setUser = user => dispatch => {
-  dispatch(fetchUsersSuccess(user));
 };
