+++ src/actions/types.js
@@ -1,21 +1,6 @@
-export const FETCH_USERS_REQUEST = 'FETCH_USERS_REQUEST';
-export const FETCH_USERS_SUCCESS = 'FETCH_USERS_SUCCESS';
-export const FETCH_USERS_FAILURE = 'FETCH_USERS_FAILURE';
-export const CREATE_USERS_REQUEST = 'CREATE_USERS_REQUEST';
-export const CREATE_USERS_SUCCESS = 'CREATE_USERS_SUCCESS';
-export const CREATE_USERS_FAILURE = 'CREATE_USERS_FAILURE';
 export const FETCH_MOTOS_REQUEST = 'FETCH_MOTOS_REQUEST';
 export const FETCH_MOTOS_SUCCESS = 'FETCH_MOTOS_SUCCESS';
 export const FETCH_MOTOS_FAILURE = 'FETCH_MOTOS_FAILURE';
-export const FETCH_FAVOURITES_REQUEST = 'FETCH_FAVOURITES_REQUEST';
-export const FETCH_FAVOURITES_SUCCESS = 'FETCH_FAVOURITES_SUCCESS';
-export const FETCH_FAVOURITES_FAILURE = 'FETCH_FAVOURITES_FAILURE';
-export const CREATE_FAVOURITE_REQUEST = 'CREATE_FAVOURITE_REQUEST';
-export const CREATE_FAVOURITE_SUCCESS = 'CREATE_FAVOURITE_SUCCESS';
-export const CREATE_FAVOURITE_FAILURE = 'CREATE_FAVOURITE_FAILURE';
-export const DELETE_FAVOURITE_REQUEST = 'DELETE_FAVOURITE_REQUEST';
-export const DELETE_FAVOURITE_SUCCESS = 'DELETE_FAVOURITE_SUCCESS';
-export const DELETE_FAVOURITE_FAILURE = 'DELETE_FAVOURITE_FAILURE';
 export const FETCH_BOOKMOTO_REQUEST = 'FETCH_BOOKMOTO_REQUEST';
 export const FETCH_BOOKMOTO_SUCCESS = 'FETCH_BOOKMOTO_SUCCESS';
 export const FETCH_BOOKMOTO_FAILURE = 'FETCH_BOOKMOTO_FAILURE';
