+++ src/components/FavouriteCard.js
@@ -2,20 +2,12 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import {
-  deleteFavourite,
-} from '../actions/index';
 import './MotoCard.css';
 
 const FavouriteCard = ({
-  id, image, name, model, deleteFavourite,
+  id, image, name, model,
 }) => {
   const rou = `/motorcycles/${id}`;
-
-  const handleRemove = e => {
-    e.preventDefault();
-    deleteFavourite(id * 1);
-  };
 
   return (
 
@@ -27,9 +19,7 @@ const FavouriteCard = ({
         <h5>{name}</h5>
         <p className="description">{model}</p>
       </Link>
-      <div className="iconsC">
-        <button type="button" className="removeButton" onClick={handleRemove}>Remove</button>
-      </div>
+
     </div>
 
   );
@@ -40,16 +30,10 @@ FavouriteCard.propTypes = {
   image: PropTypes.string.isRequired,
   name: PropTypes.string.isRequired,
   model: PropTypes.string.isRequired,
-  deleteFavourite: PropTypes.func.isRequired,
 };
-
-
-const mapDispatchToProps = dispatch => ({
-  deleteFavourite: (motoid) => dispatch(deleteFavourite(motoid)),
-});
 
 const mapStateToProps = state => ({
   favourites: state.favourites,
 });
 
-export default connect(mapStateToProps, mapDispatchToProps)(FavouriteCard);
+export default connect(mapStateToProps, null)(FavouriteCard);
