+++ src/containers/App.js
@@ -2,7 +2,6 @@ import { BrowserRouter } from 'react-router-dom';
 import Routes from './Routes';
 import './App.css';
 
-
 const App = () => (
   <BrowserRouter>
     <div className="App" data-testid="website_name">
