+++ src/index.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
+import axios from 'axios';
 import App from './containers/App';
 import store from './reducers/index';
 import './index.css';
-import axios from 'axios';
 
 axios.defaults.baseURL = 'https://motorcycles-api.herokuapp.com/api/v1';
 axios.defaults.headers.common.Authorization = `Bearer ${localStorage.getItem('motoToken')}`;
