+++ src/reducers/favourites.js
@@ -5,10 +5,8 @@ import {
   CREATE_FAVOURITE_FAILURE,
   CREATE_FAVOURITE_REQUEST,
   CREATE_FAVOURITE_SUCCESS,
-  DELETE_FAVOURITE_FAILURE,
-  DELETE_FAVOURITE_REQUEST,
-  DELETE_FAVOURITE_SUCCESS,
-} from '../actions/types';
+
+} from '../actions/favouriteTypes';
 
 const initialState = {
   motos: [],
@@ -56,25 +54,6 @@ const favouritesReducer = (state = initialState, action) => {
         loading: false,
         error: action.payload,
       };
-    case DELETE_FAVOURITE_REQUEST:
-      return {
-        ...state,
-        loading: true,
-      };
-    case DELETE_FAVOURITE_SUCCESS:
-      return {
-        ...state,
-        loading: false,
-        motos: action.payload,
-        error: '',
-      };
-    case DELETE_FAVOURITE_FAILURE:
-      return {
-        ...state,
-        loading: false,
-        error: action.payload,
-      };
-
     default:
       return state;
   }
